refactor: CategorySection 관심사 분리
