+++ frontend/src/components/Category/Header/isValidCategoryName.ts
