+++ frontend/src/components/Category/WritingList/WritingList.tsx
@@ -2,7 +2,7 @@ import { WritingIcon } from 'assets/icons';
 import { usePageNavigate } from 'hooks/usePageNavigate';
 import { useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
-import { useWritings } from '../CategorySection/useWritings';
+import { useWritings } from './useWritings';
 
 type Props = {
   categoryId: number;
